+++ src/datadoc_editor/frontend/callbacks/__init__.py
@@ -9,7 +9,9 @@ The functions where the logic lives should be categorised into files.
 
 from .register_callbacks import register_callbacks
 from .pseudonymization_callbacks import register_pseudonymization_callbacks
+from .use_restrictions_callbacks import register_use_restriction_callbacks
 
 def register_all_callbacks(app):
     register_callbacks(app)
     register_pseudonymization_callbacks(app)
+    register_use_restriction_callbacks(app)
