+++ src/datadoc_editor/frontend/callbacks/register_callbacks.py
@@ -37,21 +37,11 @@ from datadoc_editor.frontend.callbacks.variables import (
     accept_variable_metadata_date_input,
 )
 from datadoc_editor.frontend.callbacks.variables import accept_variable_metadata_input
-from datadoc_editor.frontend.callbacks.variables import mutate_variable_pseudonymization
-from datadoc_editor.frontend.callbacks.variables import populate_pseudo_workspace
 from datadoc_editor.frontend.callbacks.variables import populate_variables_workspace
 from datadoc_editor.frontend.components.builders import build_dataset_edit_section
 from datadoc_editor.frontend.components.builders import build_dataset_machine_section
 from datadoc_editor.frontend.components.identifiers import ACCORDION_WRAPPER_ID
-from datadoc_editor.frontend.components.identifiers import ADD_USE_RESTRICTION_BUTTON
-from datadoc_editor.frontend.components.identifiers import FORCE_RERENDER_COUNTER
 from datadoc_editor.frontend.components.identifiers import SECTION_WRAPPER_ID
-from datadoc_editor.frontend.components.identifiers import USE_RESTRICTION_ID_STORE
-from datadoc_editor.frontend.components.identifiers import (
-    USE_RESTRICTION_LIST_CONTAINER,
-)
-from datadoc_editor.frontend.components.identifiers import USE_RESTRICTION_OPTION_STORE
-from datadoc_editor.frontend.components.identifiers import USE_RESTRICTION_STORE
 from datadoc_editor.frontend.components.identifiers import VARIABLES_INFORMATION_ID
 from datadoc_editor.frontend.fields.display_base import DATASET_METADATA_DATE_INPUT
 from datadoc_editor.frontend.fields.display_base import DATASET_METADATA_INPUT
@@ -78,7 +68,6 @@ from datadoc_editor.frontend.fields.display_dataset import DatasetIdentifiers
 from datadoc_editor.frontend.fields.display_variables import VariableIdentifiers
 
 if TYPE_CHECKING:
-    from collections.abc import Callable
 
     import dash_bootstrap_components as dbc
 
@@ -285,95 +274,95 @@ def register_callbacks(app: Dash) -> None:  # noqa: PLR0915
             dataset_opened_counter,
         )
 
-    @app.callback(
-        [Output(USE_RESTRICTION_STORE, "data"), Output(FORCE_RERENDER_COUNTER, "data")],
-        [
-            Input(ADD_USE_RESTRICTION_BUTTON, "n_clicks"),
-            Input(
-                {
-                    "type": DATASET_METADATA_MULTIDROPDOWN_INPUT,
-                    "id": ALL,
-                    "index": ALL,
-                    "field": "type",
-                },
-                "value",
-            ),
-            Input(
-                {
-                    "type": DATASET_METADATA_MULTIDROPDOWN_INPUT,
-                    "id": ALL,
-                    "index": ALL,
-                    "field": "date",
-                },
-                "value",
-            ),
-            Input(
-                {
-                    "type": DATASET_METADATA_MULTIDROPDOWN_INPUT,
-                    "id": ALL,
-                    "index": ALL,
-                    "field": "delete",
-                },
-                "n_clicks",
-            ),
-        ],
-        [State(USE_RESTRICTION_STORE, "data"), State(FORCE_RERENDER_COUNTER, "data")],
-        prevent_initial_call=True,
-    )
-    def handle_add_and_delete(  # noqa: PLR0913
-        add_clicks: int,  # noqa: ARG001
-        type_values: list[str],
-        date_values: list[str],
-        delete_clicks: list[int],  # noqa: ARG001
-        store_data: list | None,
-        counter: int,
-    ) -> tuple[list, int]:
-        triggered = ctx.triggered_id
-        counter += 1
-        store_data = update_store_data_with_inputs(
-            store_data or [], type_values, date_values
-        )
+    #@app.callback(
+    #    [Output(USE_RESTRICTION_STORE, "data"), Output(FORCE_RERENDER_COUNTER, "data")],
+    #    [
+    #        Input(ADD_USE_RESTRICTION_BUTTON, "n_clicks"),
+    #        Input(
+    #            {
+    #                "type": DATASET_METADATA_MULTIDROPDOWN_INPUT,
+    #                "id": ALL,
+    #                "index": ALL,
+    #                "field": "type",
+    #            },
+    #            "value",
+    #        ),
+    #        Input(
+    #            {
+    #                "type": DATASET_METADATA_MULTIDROPDOWN_INPUT,
+    #                "id": ALL,
+    #                "index": ALL,
+    #                "field": "date",
+    #            },
+    #            "value",
+    #        ),
+    #        Input(
+    #            {
+    #                "type": DATASET_METADATA_MULTIDROPDOWN_INPUT,
+    #                "id": ALL,
+    #                "index": ALL,
+    #                "field": "delete",
+    #            },
+    #            "n_clicks",
+    #        ),
+    #    ],
+    #    [State(USE_RESTRICTION_STORE, "data"), State(FORCE_RERENDER_COUNTER, "data")],
+    #    prevent_initial_call=True,
+    #)
+    #def handle_add_and_delete(  # noqa: PLR0913
+    #    add_clicks: int,  # noqa: ARG001
+    #    type_values: list[str],
+    #    date_values: list[str],
+    #    delete_clicks: list[int],  # noqa: ARG001
+    #    store_data: list | None,
+    #    counter: int,
+    #) -> tuple[list, int]:
+    #    triggered = ctx.triggered_id
+    #    counter += 1
+    #    store_data = update_store_data_with_inputs(
+    #        store_data or [], type_values, date_values
+    #    )
+#
+    #    if triggered == ADD_USE_RESTRICTION_BUTTON:
+    #        store_data.append(
+    #            {"use_restriction_type": None, "use_restriction_date": None}
+    #        )
+#
+    #    elif isinstance(triggered, dict) and triggered.get("field") == "delete":
+    #        idx = triggered.get("index")
+    #        if isinstance(idx, int) and 0 <= idx < len(store_data):
+    #            remove_dataset_multidropdown_input(
+    #                DATASET_METADATA_MULTIDROPDOWN_INPUT, idx
+    #            )
+    #            store_data.pop(idx)
+#
+    #    return store_data, counter
 
-        if triggered == ADD_USE_RESTRICTION_BUTTON:
-            store_data.append(
-                {"use_restriction_type": None, "use_restriction_date": None}
-            )
-
-        elif isinstance(triggered, dict) and triggered.get("field") == "delete":
-            idx = triggered.get("index")
-            if isinstance(idx, int) and 0 <= idx < len(store_data):
-                remove_dataset_multidropdown_input(
-                    DATASET_METADATA_MULTIDROPDOWN_INPUT, idx
-                )
-                store_data.pop(idx)
-
-        return store_data, counter
-
-    @app.callback(
-        Output(USE_RESTRICTION_LIST_CONTAINER, "children"),
-        Input(USE_RESTRICTION_STORE, "data"),
-        Input(USE_RESTRICTION_OPTION_STORE, "data"),
-        Input(USE_RESTRICTION_ID_STORE, "data"),
-        Input(FORCE_RERENDER_COUNTER, "data"),
-    )
-    def render_use_restriction_list(
-        current_list: list,
-        options: Callable[[], list[dict[str, str]]],
-        idx: dict[str, str | int],
-        rerender_counter: int,
-    ) -> list:
-        if not current_list:
-            return []
-
-        items = []
-        for i, item in enumerate(current_list):
-            row_id = {**idx, "index": i}
-            items.append(
-                render_multidropdown_row(
-                    item, row_id, options, key=f"{rerender_counter}-{i}"
-                )
-            )
-        return items
+    #@app.callback(
+    #    Output(USE_RESTRICTION_LIST_CONTAINER, "children"),
+    #    Input(USE_RESTRICTION_STORE, "data"),
+    #    Input(USE_RESTRICTION_OPTION_STORE, "data"),
+    #    Input(USE_RESTRICTION_ID_STORE, "data"),
+    #    Input(FORCE_RERENDER_COUNTER, "data"),
+    #)
+    #def render_use_restriction_list(
+    #    current_list: list,
+    #    options: Callable[[], list[dict[str, str]]],
+    #    idx: dict[str, str | int],
+    #    rerender_counter: int,
+    #) -> list:
+    #    if not current_list:
+    #        return []
+#
+    #    items = []
+    #    for i, item in enumerate(current_list):
+    #        row_id = {**idx, "index": i}
+    #        items.append(
+    #            render_multidropdown_row(
+    #                item, row_id, options, key=f"{rerender_counter}-{i}"
+    #            )
+    #        )
+    #    return items
 
     @app.callback(
         Output(SECTION_WRAPPER_ID, "children"),
